pyproject.toml
[tool.poetry]
name = "screenshot-bot"
version = "0.1.0"
description = "Simple Telegram bot to capture a screenshot of selected page"
authors = ["Filinovych Danylo <legio9116@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^13.0"
envparse = "^0.2.0"
pyppeteer = "^0.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
